+++ capstone_project/src/components/CommentList.jsx
@@ -1,19 +1,20 @@
 import { useEffect, useState } from "react";
 
-const CommentList = () => {
+const CommentList = ({currentComments}) => {
   //this will be a list of comments
-  const [items, setItems] = useState(['Item 1', 'Item 2', 'Item 3']);
 
-  const handleClick = (item) => {
-    console.log(`You have clicked on: ${item}`);
+  console.log(currentComments);
+
+  const handleClick = (thisComment) => {
+    console.log(`You have clicked on: ${thisComment.Name} and here is the time: ${thisComment.Response_Date_Time__c}`);
   }
 
     return (
       <>
         <h4> Here is the comment list </h4>
         <ul>
-          {items.length && items.map((item, index) => (
-            <li key={index} onClick={() => handleClick(item)}>{item}</li>
+          {currentComments.length && currentComments.map((thisComment) => (
+            <li key={thisComment.Id} onClick={() => handleClick(thisComment)}>{thisComment.Client_Message__c}</li>
           ))}
         </ul>
       </>
