+++ capstone_project/src/components/register.jsx
@@ -1,5 +1,7 @@
 /* TODO - add your code to create a functional React component that renders a registration form */
 import { useEffect, useState } from "react";
+import { useParams, useNavigate} from "react-router-dom";
+import { getSalesforceAccessToken, registerUser } from "../api/api";
 //import getSalesforceAccessToken from "../api/api";
 //import dotenv from 'dotenv';
 import axios from "axios";
@@ -8,92 +10,94 @@ import axios from "axios";
 
 //hardcoded stuff
 
-async function getSalesforceAccessToken() {
-  const tokenUrl = `https://interaudibank-dev-ed.develop.my.salesforce.com/services/oauth2/token`;
-  try {
-    console.log("process.env", process.env);
-    console.log(
-      "process.env.REACT_APP_SALESFORCE_PASSWORD",
-      process.env.REACT_APP_SALESFORCE_PASSWORD
-    );
-    console.log(
-      "process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN",
-      process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN
-    );
-    const response = await axios.post(tokenUrl, null, {
-      params: {
-        grant_type: "password", // This example uses Username-Password OAuth flow
-        client_id: process.env.REACT_APP_CLIENT_ID,
-        client_secret: process.env.REACT_APP_CLIENT_SECRET,
-        username: process.env.REACT_APP_SALESFORCE_USERNAME,
-        password:
-          process.env.REACT_APP_SALESFORCE_PASSWORD +
-          process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN,
-      },
-    });
+// async function getSalesforceAccessToken() {
+//   const tokenUrl = `https://interaudibank-dev-ed.develop.my.salesforce.com/services/oauth2/token`;
+//   try {
+//     console.log("process.env", process.env);
+//     console.log(
+//       "process.env.REACT_APP_SALESFORCE_PASSWORD",
+//       process.env.REACT_APP_SALESFORCE_PASSWORD
+//     );
+//     console.log(
+//       "process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN",
+//       process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN
+//     );
+//     const response = await axios.post(tokenUrl, null, {
+//       params: {
+//         grant_type: "password", // This example uses Username-Password OAuth flow
+//         client_id: process.env.REACT_APP_CLIENT_ID,
+//         client_secret: process.env.REACT_APP_CLIENT_SECRET,
+//         username: process.env.REACT_APP_SALESFORCE_USERNAME,
+//         password:
+//           process.env.REACT_APP_SALESFORCE_PASSWORD +
+//           process.env.REACT_APP_SALESFORCE_SECURITY_TOKEN,
+//       },
+//     });
 
-    console.log(response.data.access_token);
-    return response.data.access_token;
-  } catch (error) {
-    console.error("Error getting Salesforce access token:", error);
-    throw error;
-  }
-}
+//     console.log(response.data.access_token);
+//     return response.data.access_token;
+//   } catch (error) {
+//     console.error("Error getting Salesforce access token:", error);
+//     throw error;
+//   }
+// }
 
-const Register = ({ setToken }) => {
+const Register = ({ setDocumentList }) => {
   const [selectedFile, setSelectedFile] = useState(null);
+  const navigate = useNavigate();
 
-  async function registerUser(e) {
-    e.preventDefault();
-    let firstName = document.getElementById("firstname").value;
-    let lastName = document.getElementById("lastname").value;
-    let emailValue = document.getElementById("email").value;
-    let addressValue = document.getElementById("address").value;
-    let passwordValue = document.getElementById("password").value;
-    let usernameValue = document.getElementById("username").value;
-    let phoneNumberValue = document.getElementById("phoneNumber").value;
-    let accountTypeValue = document.getElementById("accountType").value;
+  // async function registerUser(e) {
+  //   e.preventDefault();
+  //   let firstName = document.getElementById("firstname").value;
+  //   let lastName = document.getElementById("lastname").value;
+  //   let emailValue = document.getElementById("email").value;
+  //   let addressValue = document.getElementById("address").value;
+  //   let passwordValue = document.getElementById("password").value;
+  //   let usernameValue = document.getElementById("username").value;
+  //   let phoneNumberValue = document.getElementById("phoneNumber").value;
+  //   let accountTypeValue = document.getElementById("accountType").value;
 
-    console.log(firstName);
-    console.log(lastName);
-    console.log(emailValue);
-    console.log(addressValue);
-    console.log(usernameValue);
-    console.log(passwordValue);
-    console.log(phoneNumberValue);
-    console.log(accountTypeValue);
+  //   console.log(firstName);
+  //   console.log(lastName);
+  //   console.log(emailValue);
+  //   console.log(addressValue);
+  //   console.log(usernameValue);
+  //   console.log(passwordValue);
+  //   console.log(phoneNumberValue);
+  //   console.log(accountTypeValue);
 
-    try {
-      const accessToken = await getSalesforceAccessToken();
-      console.log("accessToken", accessToken);
-      const response = await fetch(
-        "https://interaudibank-dev-ed.develop.my.salesforce.com/services/apexrest/api/Account_Opening/Register",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${accessToken}`,
-          },
-          body: JSON.stringify({
-            firstName: firstName,
-            lastName: lastName,
-            email: emailValue,
-            address: addressValue,
-            username: usernameValue,
-            password: passwordValue,
-            typeofBankAccount: accountTypeValue,
-            phoneNumber: phoneNumberValue,
-          }),
-        }
-      );
-      const result = await response.json();
-      console.log(result);
-      console.log(result.token);
-      setToken(result.token);
-    } catch (e) {
-      console.log(e);
-    }
-  }
+  //   try {
+  //     const accessToken = await getSalesforceAccessToken();
+  //     console.log("accessToken", accessToken);
+  //     const response = await fetch(
+  //       "https://interaudibank-dev-ed.develop.my.salesforce.com/services/apexrest/api/Account_Opening/Register",
+  //       {
+  //         method: "POST",
+  //         headers: {
+  //           "Content-Type": "application/json",
+  //           Authorization: `Bearer ${accessToken}`,
+  //         },
+  //         body: JSON.stringify({
+  //           firstName: firstName,
+  //           lastName: lastName,
+  //           email: emailValue,
+  //           address: addressValue,
+  //           username: usernameValue,
+  //           password: passwordValue,
+  //           typeofBankAccount: accountTypeValue,
+  //           phoneNumber: phoneNumberValue,
+  //         }),
+  //       }
+  //     );
+  //     const result = await response.json();
+  //     console.log(result);
+  //     //console.log(result.token);
+  //     //setToken(result.token);
+  //     navigate('/username')
+  //   } catch (e) {
+  //     console.log(e);
+  //   }
+  // }
 
   async function loginUser(e) {
     /*
@@ -120,34 +124,34 @@ let usernameValue = document.getElementById("userNameLogin").value;
       const result = await response.json();
       console.log(result);
       console.log(result.token);
-      setToken(result.token);
+      //setToken(result.token);
     } catch (e) {
       console.log(e);
     }
   }
 
-  async function getClientDocs(e) {
-    try {
-      const accessToken = await getSalesforceAccessToken();
-      console.log("accessToken", accessToken);
-      const response = await fetch(
-        "https://interaudibank-dev-ed.develop.my.salesforce.com/services/apexrest/api/Client_Documents/a00aj00000MzQAb",
-        {
-          method: "GET",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${accessToken}`,
-          },
-        }
-      );
-      const result = await response.json();
-      console.log(result);
-      console.log(result.token);
-      setToken(result.token);
-    } catch (e) {
-      console.log(e);
-    }
-  }
+  // async function getClientDocs(e) {
+  //   try {
+  //     const accessToken = await getSalesforceAccessToken();
+  //     console.log("accessToken", accessToken);
+  //     const response = await fetch(
+  //       "https://interaudibank-dev-ed.develop.my.salesforce.com/services/apexrest/api/Client_Documents/a00aj00000MzQAb",
+  //       {
+  //         method: "GET",
+  //         headers: {
+  //           "Content-Type": "application/json",
+  //           Authorization: `Bearer ${accessToken}`,
+  //         },
+  //       }
+  //     );
+  //     const result = await response.json();
+  //     console.log(result);
+  //     console.log(result.token);
+  //     //setToken(result.token);
+  //   } catch (e) {
+  //     console.log(e);
+  //   }
+  // }
 
   async function getClientComments(e) {
     try {
@@ -166,7 +170,7 @@ let usernameValue = document.getElementById("userNameLogin").value;
       const result = await response.json();
       console.log(result);
       console.log(result.token);
-      setToken(result.token);
+      //setToken(result.token);
     } catch (e) {
       console.log(e);
     }
@@ -193,7 +197,7 @@ let usernameValue = document.getElementById("userNameLogin").value;
       const result = await response.json();
       console.log(result);
       console.log(result.token);
-      setToken(result.token);
+      //setToken(result.token);
     } catch (e) {
       console.log(e);
     }
@@ -216,7 +220,7 @@ let usernameValue = document.getElementById("userNameLogin").value;
       const result = await response.json();
       console.log(result);
       console.log(result.token);
-      setToken(result.token);
+      //setToken(result.token);
     } catch (e) {
       console.log(e);
     }
@@ -268,7 +272,7 @@ let usernameValue = document.getElementById("userNameLogin").value;
       const result = await response.json();
       console.log(result);
       console.log(result.token);
-      setToken(result.token);
+      //setToken(result.token);
     } catch (e) {
       console.log(e);
     }
@@ -321,7 +325,7 @@ let usernameValue = document.getElementById("userNameLogin").value;
 
         <br></br>
         <br></br>
-        <button onClick={(e) => registerUser(e)}>Register</button>
+        <button onClick={async (e) => {await registerUser(e, setDocumentList); navigate('/username')}}>Register</button>
 
         <br></br>
         <br></br>
